+++ newpage.html
@@ -4,12 +4,18 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>hello world</title>
+  <title>新的页面</title>
 </head>
 <body>
-  <button id="btn">主页面</button>
-  <div id="baby"></div>
-  <button id="newpage">新的页面</button>
+<div class="new">
+  <p>新的页面</p>
+</div>
 </body>
-<script src="./render/index.js"></script>
+<style>
+  .new {
+    width: 100px;
+    height: 100px;
+    background-color: red;
+  }
+</style>
 </html>
