+++ package.json
@@ -8,77 +8,78 @@
   "author": "xiaxin83",
   "license": "ISC",
   "devDependencies": {
+    "@electron/remote": "^2.0.8",
     "electron": "^21.2.0"
   },
   "dependencies": {
     "boolean": "^3.2.0",
-    "config-chain": "^1.1.13",
+    "buffer-crc32": "^0.2.13",
     "cacheable-request": "^6.1.0",
+    "clone-response": "^1.0.3",
+    "config-chain": "^1.1.13",
     "debug": "^4.3.4",
     "decompress-response": "^3.3.0",
-    "buffer-crc32": "^0.2.13",
-    "clone-response": "^1.0.3",
-    "encodeurl": "^1.0.2",
     "defer-to-connect": "^1.1.3",
     "define-properties": "^1.1.4",
     "detect-node": "^2.1.0",
+    "duplexer3": "^0.1.5",
+    "encodeurl": "^1.0.2",
     "end-of-stream": "^1.4.4",
     "env-paths": "^2.2.1",
     "es6-error": "^4.1.1",
     "escape-string-regexp": "^4.0.0",
     "extract-zip": "^2.0.1",
+    "fd-slicer": "^1.1.0",
     "fs-extra": "^8.1.0",
     "function-bind": "^1.1.1",
-    "duplexer3": "^0.1.5",
     "get-intrinsic": "^1.1.3",
-    "has": "^1.0.3",
+    "get-stream": "^5.2.0",
     "global-agent": "^3.0.0",
     "global-tunnel-ng": "^2.7.1",
-    "fd-slicer": "^1.1.0",
     "globalthis": "^1.0.3",
-    "ini": "^1.3.8",
-    "get-stream": "^5.2.0",
-    "http-cache-semantics": "^4.1.0",
+    "got": "^9.6.0",
+    "graceful-fs": "^4.2.10",
+    "has": "^1.0.3",
     "has-property-descriptors": "^1.0.0",
-    "json-stringify-safe": "^5.0.1",
+    "has-symbols": "^1.0.3",
+    "http-cache-semantics": "^4.1.0",
+    "ini": "^1.3.8",
     "json-buffer": "^3.0.0",
+    "json-stringify-safe": "^5.0.1",
     "jsonfile": "^4.0.0",
+    "keyv": "^3.1.0",
     "lodash": "^4.17.21",
     "lowercase-keys": "^1.0.1",
-    "has-symbols": "^1.0.3",
-    "keyv": "^3.1.0",
-    "matcher": "^3.0.0",
-    "ms": "^2.1.2",
-    "mimic-response": "^1.0.1",
     "lru-cache": "^6.0.0",
-    "npm-conf": "^1.1.3",
+    "matcher": "^3.0.0",
+    "mimic-response": "^1.0.1",
+    "ms": "^2.1.2",
     "normalize-url": "^4.5.1",
-    "p-cancelable": "^1.1.0",
-    "once": "^1.4.0",
-    "pend": "^1.2.0",
-    "proto-list": "^1.2.4",
+    "npm-conf": "^1.1.3",
     "object-keys": "^1.1.1",
-    "pump": "^3.0.0",
+    "once": "^1.4.0",
+    "p-cancelable": "^1.1.0",
+    "pend": "^1.2.0",
     "pify": "^3.0.0",
     "prepend-http": "^2.0.0",
-    "semver": "^6.3.0",
-    "roarr": "^2.15.4",
+    "progress": "^2.0.3",
+    "proto-list": "^1.2.4",
+    "pump": "^3.0.0",
     "responselike": "^1.0.2",
+    "roarr": "^2.15.4",
+    "semver": "^6.3.0",
     "semver-compare": "^1.0.0",
-    "sprintf-js": "^1.1.2",
     "serialize-error": "^7.0.1",
+    "sprintf-js": "^1.1.2",
+    "sumchecker": "^3.0.1",
+    "to-readable-stream": "^1.0.0",
     "tunnel": "^0.0.6",
     "type-fest": "^0.13.1",
-    "to-readable-stream": "^1.0.0",
+    "universalify": "^0.1.2",
     "url-parse-lax": "^3.0.0",
     "wrappy": "^1.0.2",
-    "universalify": "^0.1.2",
-    "sumchecker": "^3.0.1",
     "yallist": "^4.0.0",
-    "yauzl": "^2.10.0",
-    "progress": "^2.0.3",
-    "got": "^9.6.0",
-    "graceful-fs": "^4.2.10"
+    "yauzl": "^2.10.0"
   },
   "repository": {
     "type": "git",
